[refactor] service command/mapper 패키지 구조 개선 - #17
