+++ novelservice/src/main/java/com/iucyh/novelservice/novel/service/NovelService.java
@@ -5,16 +5,12 @@ import com.iucyh.novelservice.novel.exception.DuplicateNovelTitle;
 import com.iucyh.novelservice.novel.exception.HasNoEpisodes;
 import com.iucyh.novelservice.novel.exception.NovelNotFound;
 import com.iucyh.novelservice.novel.domain.Novel;
-import com.iucyh.novelservice.novel.enumtype.NovelCategory;
-import com.iucyh.novelservice.novel.service.command.CreateNovelCommand;
-import com.iucyh.novelservice.novel.service.command.UpdateNovelCommand;
-import com.iucyh.novelservice.novel.service.command.mapper.NovelCommandMapper;
-import com.iucyh.novelservice.novel.web.dto.mapper.NovelRequestMapper;
+import com.iucyh.novelservice.novel.service.dto.command.CreateNovelCommand;
+import com.iucyh.novelservice.novel.service.dto.command.UpdateNovelCommand;
+import com.iucyh.novelservice.novel.service.dto.mapper.NovelCommandMapper;
 import com.iucyh.novelservice.novel.web.dto.mapper.NovelResponseMapper;
-import com.iucyh.novelservice.novel.web.dto.request.CreateNovelRequest;
 import com.iucyh.novelservice.novel.web.dto.response.NovelCompletionResponse;
 import com.iucyh.novelservice.novel.web.dto.response.NovelLikeCountResponse;
-import com.iucyh.novelservice.novel.web.dto.request.UpdateNovelRequest;
 import com.iucyh.novelservice.novel.web.dto.response.NovelResponse;
 import com.iucyh.novelservice.novel.repository.NovelRepository;
 import com.iucyh.novelservice.user.domain.User;
